generate-report-lambda/pyproject.toml
[project]
name = "generate-report"
version = "0.1.0"
description = "A Lambda function for generating reports"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.35.92",
    "docxtpl>=0.19.1",
    "matplotlib>=3.9.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
